Fix: Standardize local server architecture to match other servers (#2)
Consolidate multiple entry points into single main.py following MCP Fleet
architecture pattern used by compass, tides, and toolkit servers.

Changes:
- Remove mcp_main.py (consolidated into main.py)
- Move run_server.py to tests/standalone_test_server.py
- Update main.py to use direct mcp imports consistently
- Remove sys.path manipulation in favor of standard pattern
- Update Claude Desktop config to use servers/local/main.py
- Maintain all Apple Notes functionality

Architecture now matches other servers:
- Single main.py entry point
- Consistent MCP protocol implementation
- Standard import patterns
- Proper tool registration and handlers

Fixes #2
